fix: make lint, types and tests pass; fix a real preprocessing bug
First run of the toolchain against this code. Three gates now pass locally:
ruff check + format clean, mypy strict clean over 31 files, 165 tests passing.

The substantive fix is in app/serving.py. Its resize computed the long side
with round() while torchvision truncates, so a 400x300 photo resized to a
299px long side instead of 298. That one pixel shifts the subsequent centre
crop, producing an entirely different tensor -- the deployed demo would have
preprocessed most non-square photographs differently from training, and the
demo's accuracy would not have matched the reported accuracy.

It was mostly latent: with the 256px resize the two agree by coincidence for
common aspect ratios, so only the CLIP spec (resize 224) exposed it.
test_serving_parity.py caught it on the first run, which is the reason that
test exists.

Also fixed:
- Image.BILINEAR is a legacy alias no longer in Pillow's stubs; use
  Image.Resampling.BILINEAR.
- Redundant int(round(...)) casts -- round() with one argument already
  returns an int.
- An en dash inside an f-string, flagged as an ambiguous character.
- Import ordering, formatting, and over-length lines across 18 files.
- mypy strict: bind nn.Module call results to typed locals rather than
  returning Any; cast CLIP's encode_image, which nn.Module.__getattr__ types
  as Tensor; parse manifest values through str() instead of suppressing;
  narrow numpy integer indices before indexing a Sequence.

+++ src/carvision/data/splits.py
@@ -87,7 +87,7 @@ def stratified_val_split(
         permuted = rng.permutation(indices)
 
         # At least one validation image per class, and at least one left in train.
-        n_val = int(round(len(indices) * val_fraction))
+        n_val = round(len(indices) * val_fraction)
         n_val = max(1, min(n_val, len(indices) - 1))
         mask[permuted[:n_val]] = True
 
@@ -118,9 +118,7 @@ def build_splits(
 
     manifest_path = data_dir() / "stanford_cars" / "manifest.csv"
     if not manifest_path.exists():
-        raise SplitError(
-            f"{manifest_path} not found. Run `carvision data download` first."
-        )
+        raise SplitError(f"{manifest_path} not found. Run `carvision data download` first.")
 
     out_dir = ensure_dir(splits_dir())
     existing = [p for p in (out_dir / f"{s}.csv" for s in SPLIT_NAMES) if p.exists()]
